(refactor) Move nearby map to own component
